euron_rag_application/requirements.txt
streamlit
faiss-cpu
numpy
requests
python-env
